refactor(codex-manager): extract standalone formatters into formatters/
26 standalone, side-effect-free formatter/presentation functions (plus the
PromptTone and ModelInspection types) moved verbatim from lib/codex-manager.ts
into lib/codex-manager/formatters/{text-style,quota-formatters,
account-formatters,model-formatters}.ts with an index.ts barrel.
codex-manager.ts shrinks from 3810 to 3446 lines; the previously public
exports styleAccountDetailText and formatBackupSavedAt are re-exported from
lib/codex-manager.ts so every existing import path keeps working; zero
behavior change. Phase 1 of audit roadmap §4.1.1.
